+++ asteroid-core/src/main/java/asteroid/local/LocalTransformation.java
@@ -2,11 +2,12 @@ package asteroid.local;
 
 import org.codehaus.groovy.transform.GroovyASTTransformationClass;
 
-import asteroid.A;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
+import org.codehaus.groovy.control.CompilePhase;
 
 /**
  * Can be used when declaring an implementation of {@link LocalTransformationImpl} to reduce
@@ -30,5 +31,5 @@ public @interface LocalTransformation {
      * @return the specific {@link asteroid.A.PHASE_LOCAL} when transformation will be applied
      * @since 0.1.0
      */
-    A.PHASE_LOCAL value();
+    CompilePhase value();
 }