+++ asteroid-core/src/main/java/asteroid/utils/ClassNodeUtils.java
@@ -12,8 +12,11 @@ import org.codehaus.groovy.ast.MethodNode;
 import org.codehaus.groovy.ast.ModuleNode;
 import org.codehaus.groovy.ast.PropertyNode;
 import org.codehaus.groovy.ast.ClassHelper;
+import org.codehaus.groovy.ast.AnnotatedNode;
 import org.codehaus.groovy.ast.AnnotationNode;
 import org.codehaus.groovy.ast.tools.GeneralUtils;
+
+import org.codehaus.groovy.control.CompilePhase;
 
 import groovy.lang.Closure;
 
@@ -78,8 +81,12 @@ public final class ClassNodeUtils {
     /**
      * Makes the {@link ClassNode} to implement the interfaces passed
      * as arguments
+     * <br><br>
+     * <b>IMPORTANT</b>: Dont use this method at any {@link CompilePhase}
+     * before SEMANTIC_ANALYSIS. Classes may have not been set at this
+     * point.
      *
-     * @param classNode
+     * @param classNode The class we want to add the interfaces to
      * @param interfaces the interfaces we want the class node to be
      * implementing
      * @since 0.1.4
@@ -87,6 +94,22 @@ public final class ClassNodeUtils {
     public void addInterfaces(final ClassNode classNode, final Class... interfaces) {
         for (final Class clazz : interfaces) {
             final ClassNode nextInterface = ClassHelper.make(clazz, false);
+            classNode.addInterface(nextInterface);
+        }
+    }
+
+    /**
+     * Makes the {@link ClassNode} to implement the interfaces passed
+     * as arguments.
+     *
+     * @param classNode the {@link ClassNode} we want to implement
+     * certain interfaces
+     * @param interfaces the interfaces we want our {@link ClassNode}
+     * to implement
+     * @since 0.1.7
+     */
+    public void addInterfaces(final ClassNode classNode, final ClassNode... interfaces) {
+        for (final ClassNode nextInterface : interfaces) {
             classNode.addInterface(nextInterface);
         }
     }
@@ -128,6 +151,17 @@ public final class ClassNodeUtils {
      */
     public AnnotationNode getAnnotationFrom(final ClassNode classNode, final String simpleName) {
         return find(classNode.getAnnotations(), byName(simpleName));
+    }
+
+    /**
+     * It removes the {@link AnnotationNode} from a given {@link AnnotatedNode}
+     *
+     * @param annotated the {@link AnnotatedNode} to remove the annotation from
+     * @param annotation the {@link AnnotationNode} you want to remove
+     * @since 0.1.7
+     */
+    public void removeAnnotation(final AnnotatedNode annotated, final AnnotationNode annotation) {
+        annotated.getAnnotations().remove(annotation);
     }
 
     private Closure<Boolean> byName(final String annotationName) {