+++ asteroid-test/src/main/groovy/asteroid/local/samples/ExecuteMethodSafelyImpl.groovy
@@ -8,14 +8,13 @@ import groovy.transform.CompileStatic
 import org.codehaus.groovy.ast.MethodNode
 import org.codehaus.groovy.ast.AnnotationNode
 import org.codehaus.groovy.ast.FieldNode
-import org.codehaus.groovy.control.SourceUnit
 
 @CompileStatic
 @LocalTransformation(A.PHASE_LOCAL.SEMANTIC_ANALYSIS)
 class ExecuteMethodSafelyImpl extends LocalTransformationImpl<ExecuteMethodSafely, MethodNode> {
 
     @Override
-    void doVisit(AnnotationNode annotation, MethodNode methodNode, SourceUnit source) {
+    void doVisit(AnnotationNode annotation, MethodNode methodNode) {
         List<FieldNode> fields = A.UTIL.CLASS.getInstancePropertyFields(methodNode.declaringClass)
         String valueFieldNode = fields.find()?.name