+++ asteroid-test/src/main/groovy/asteroid/local/samples/ToMD5Impl.groovy
@@ -10,14 +10,13 @@ import org.codehaus.groovy.ast.AnnotationNode
 import org.codehaus.groovy.ast.FieldNode
 import org.codehaus.groovy.ast.MethodNode
 import org.codehaus.groovy.ast.stmt.BlockStatement
-import org.codehaus.groovy.control.SourceUnit
 
 @CompileStatic
 @LocalTransformation(A.PHASE_LOCAL.SEMANTIC_ANALYSIS) // <1>
 class ToMD5Impl extends LocalTransformationImpl<ToMD5, FieldNode> { // <2>
 
     @Override
-    void doVisit(AnnotationNode annotation, FieldNode node, SourceUnit source) { // <3>
+    void doVisit(AnnotationNode annotation, FieldNode node) { // <3>
         BlockStatement block  = buildMethodCode(node.name)
         MethodNode methodNode = A.NODES.method("${node.name}ToMD5")
             .modifiers(A.ACC.ACC_PUBLIC)