+++ asteroid-test/src/test/groovy/asteroid/global/samples/AddTransformerSpecExample.groovy
@@ -1,5 +1,6 @@
 package asteroid.global.samples
 
+@TrickOrTrait(Something)
 class AddTransformerSpecExample {
 
     static class Input {
@@ -8,5 +9,15 @@ class AddTransformerSpecExample {
 
     Input execute() {
         return new Input(materialId: 1L)
+    }
+}
+
+@interface TrickOrTrait {
+    Class value();
+}
+
+trait Something {
+    String saySomething() {
+        return "something"
     }
 }